feat(rs-transport): introduce `{Serve,Invoke}Ext`
