+++ Cargo.toml
@@ -128,16 +128,16 @@ wasmtime = { version = "22", default-features = false }
 wasmtime-wasi = { version = "22", default-features = false }
 wit-bindgen = { version = "0.27", default-features = false }
 wit-bindgen-core = { version = "0.27", default-features = false }
-wit-bindgen-wrpc = { version = "0.4.5", default-features = false, path = "./crates/wit-bindgen" }
+wit-bindgen-wrpc = { version = "0.4.6", default-features = false, path = "./crates/wit-bindgen" }
 wit-bindgen-wrpc-go = { version = "0.2", default-features = false, path = "./crates/wit-bindgen-go" }
-wit-bindgen-wrpc-rust = { version = "0.4.5", default-features = false, path = "./crates/wit-bindgen-rust" }
-wit-bindgen-wrpc-rust-macro = { version = "0.4.5", default-features = false, path = "./crates/wit-bindgen-rust-macro" }
+wit-bindgen-wrpc-rust = { version = "0.4.6", default-features = false, path = "./crates/wit-bindgen-rust" }
+wit-bindgen-wrpc-rust-macro = { version = "0.4.6", default-features = false, path = "./crates/wit-bindgen-rust-macro" }
 wit-component = { version = "0.212", default-features = false }
 wit-parser = { version = "0.212", default-features = false }
 wrpc-cli = { version = "0.2", path = "./crates/cli", default-features = false }
 wrpc-introspect = { version = "0.1", default-features = false, path = "./crates/introspect" }
 wrpc-runtime-wasmtime = { version = "0.19", path = "./crates/runtime-wasmtime", default-features = false }
-wrpc-transport = { version = "0.26.3", path = "./crates/transport", default-features = false }
-wrpc-transport-nats = { version = "0.22", path = "./crates/transport-nats", default-features = false }
+wrpc-transport = { version = "0.26.4", path = "./crates/transport", default-features = false }
+wrpc-transport-nats = { version = "0.22.1", path = "./crates/transport-nats", default-features = false }
 wrpc-transport-quic = { version = "0.1", path = "./crates/transport-quic", default-features = false }
 wrpc-wasmtime-nats-cli = { version = "0.3", path = "./crates/wasmtime-nats-cli", default-features = false }
