+++ crates/transport-nats/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wrpc-transport-nats"
-version = "0.22.0"
+version = "0.22.1"
 description = "wRPC NATS transport"
 
 authors.workspace = true
