+++ crates/wit-bindgen-rust-macro/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wit-bindgen-wrpc-rust-macro"
-version = "0.4.5"
+version = "0.4.6"
 description = """
 Procedural macro paired with the `wit-bindgen-wrpc` crate.
 """
