+++ crates/wit-bindgen-rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wit-bindgen-wrpc-rust"
-version = "0.4.5"
+version = "0.4.6"
 description = """
 Rust bindings generator for wRPC, typically used through
 the `wit-bindgen-wrpc` crate's `generate!` macro.
