+++ crates/wit-bindgen/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wit-bindgen-wrpc"
-version = "0.4.5"
+version = "0.4.6"
 description = """
 Rust bindings generator for wRPC.
 """
